refactor: restructure folders following NPM guidelines
Reorganized project folder structure to clearly separate public/private API boundaries according to NPM guidelines

- src/core/ → src/pkg/pull/ (package core features)
- src/push/ → src/pkg/push/ (package core features)
- src/utils/ → src/internal/utils/ (private utilities)
- src/types/ → src/internal/types/ (private types)
- Removed src/core/providers/detector.ts (inlined as it always returns "github")

Introduced pkg/ layer to place pull/push features at the same level for consistency,
and internal/ layer to explicitly mark private APIs providing clear API boundaries to external users

fix #67
