--- src/core/providers/detector.ts
@@ -1,6 +1,0 @@
-import type { Provider } from "../../types/providers.js";
-
-export const detectProvider = (source: string): Provider => {
-  void source;
-  return "github";
-};
