+++ src/internal/utils/check-existing.ts
