+++ src/internal/utils/extract.ts
